requirements.txt
boto3==1.10.4
botocore==1.13.4
Django==2.2.6
docutils==0.15.2
jmespath==0.9.4
psycopg2-binary==2.8.4
python-dateutil==2.8.0
python-dotenv==0.10.3
pytz==2019.3
s3transfer==0.2.1
six==1.12.0
sqlparse==0.3.0
urllib3==1.25.6
Werkzeug==0.16.0
